% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAlpha2.R
\name{findAlpha2}
\alias{findAlpha2}
\title{Internal function to get second-stage level for inverse normal test}
\usage{
findAlpha2(alpha2, alpha, alpha1, alpha0, weights)
}
\arguments{
\item{alpha2}{Current value of the second-stage significance level for which solving of the level condition is attempted. Internally specified.}

\item{alpha}{The overall type I error rate \eqn{\alpha} of the design. Must be a numeric value between 0 and 1.}

\item{alpha1}{Stage 1 efficacy boundary \eqn{\alpha_1} (p-value scale). Must be a numeric value between 0 and 1. Should be smaller than \code{alpha0}.}

\item{alpha0}{Binding stage 1 futility boundary \eqn{\alpha_0} (p-value scale). Must be a numeric value between 0 and 1. Should be greater than \code{alpha1}. For use of a non-binding futility boundary, specify \code{alpha0=1}.}

\item{weights}{Weights assigned to the p-value/test statistic of each stage. Must be a numeric vector assigning a value between 0 and 1 to each stage that sums up to 1.}
}
\value{
Distance of the current value of alpha2 to the value that solves the level condition.
}
\description{
Helper function to identify second-stage significance level for the inverse normal test
}
\keyword{internal}
