% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlpha2InverseNormal.R
\name{getAlpha2InverseNormal}
\alias{getAlpha2InverseNormal}
\title{Get Second-stage Level for Inverse Normal Method}
\usage{
getAlpha2InverseNormal(alpha, alpha1, alpha0, weights = 1/sqrt(c(2, 2)))
}
\arguments{
\item{alpha}{The overall type I error rate \eqn{\alpha} of the design. Must be a numeric value between 0 and 1.}

\item{alpha1}{Stage 1 efficacy boundary \eqn{\alpha_1} (p-value scale). Must be a numeric value between 0 and 1. Should be smaller than \code{alpha0}.}

\item{alpha0}{Binding stage 1 futility boundary \eqn{\alpha_0} (p-value scale). Must be a numeric value between 0 and 1. Should be greater than \code{alpha1}. For use of a non-binding futility boundary, specify \code{alpha0=1}.}

\item{weights}{Weights assigned to the p-value/test statistic of each stage. Must be a numeric vector assigning a value between 0 and 1 to each stage that sums up to 1.}
}
\value{
Second-stage significance level of the inverse normal test.
}
\description{
Find second-stage significance level for inverse normal test (\code{getInverseNormalCombinationFunction()}) and inverse normal conditional error (\code{getInverseNormalConditionalError()}).
}
\examples{
getAlpha2InverseNormal(alpha=0.025, alpha1=0.001, alpha0=0.5)

}
