% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircularConditionalError.R
\name{getCircularConditionalError}
\alias{getCircularConditionalError}
\title{Get Circular Conditional Error}
\usage{
getCircularConditionalError(firstStagePValue, alpha1, alpha0)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{alpha1}{Stage 1 efficacy boundary \eqn{\alpha_1} (p-value scale). Must be a numeric value between 0 and 1. Should be smaller than \code{alpha0}.}

\item{alpha0}{Binding stage 1 futility boundary \eqn{\alpha_0} (p-value scale). Must be a numeric value between 0 and 1. Should be greater than \code{alpha1}. For use of a non-binding futility boundary, specify \code{alpha0=1}.}
}
\value{
Value of the circular conditional error function.
}
\description{
Calculate circular conditional error according to Proschan & Hunsberger (1995).
}
\details{
For specified early decision boundaries \eqn{\alpha_0, \alpha_1}, the circular conditional error function is calculated for a first-stage p-value \eqn{p_1 \in ]\alpha_1, \alpha_0]} as:
\deqn{\alpha_2(p_1) = 1-\Phi(\sqrt{(\Phi^{-1}(1-\alpha_1))^2 - (\Phi^{-1}(1-p_1))^2}).}
For a tabulation of early decision boundaries, refer to Wassmer & Brannath (2016).
}
\examples{
getCircularConditionalError(firstStagePValue=0.1, alpha1=0.01170, alpha0=0.5)

}
\references{
Proschan, M. A., & Hunsberger, S. A. (1995). Designed extension of studies based on conditional power. Biometrics. http://www.jstor.org/stable/2533262

Wassmer, G., & Brannath, W. (2016). Group sequential and confirmatory adaptive designs in clinical trials. Springer. https://doi.org/https://doi.org/10.1007/978-3-319-32562-0
}
