% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditionalErrorFromSecondStageInformation.R
\name{getConditionalErrorFromSecondStageInformation}
\alias{getConditionalErrorFromSecondStageInformation}
\title{Calculate conditional error from second-stage information}
\usage{
getConditionalErrorFromSecondStageInformation(
  firstStagePValue,
  secondStageInformation,
  design
)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{secondStageInformation}{The second-stage information for which to calculate the corresponding conditional error.}

\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}
}
\value{
The conditional error which corresponds to the provided second-stage information.
}
\description{
This function is used to determine the maximum and minimum conditional error resulting from specified minimum and maximum second-stage information.
}
