% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstraints.R
\name{getConstraints}
\alias{getConstraints}
\title{Calculates the constraints.}
\usage{
getConstraints(firstStagePValue, design)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the optimal conditional error function for the specified case.}
}
\value{
List with the lower and upper constraint
}
\description{
Calculates the constraints.
}
\details{
Internal function that calculates the constraints.
}
