% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDesignOptimalConditionalErrorFunction.R
\name{getDesignOptimalConditionalErrorFunction}
\alias{getDesignOptimalConditionalErrorFunction}
\title{Create a design object for the optimal conditional error function.}
\usage{
getDesignOptimalConditionalErrorFunction(
  alpha,
  alpha1,
  alpha0,
  conditionalPower,
  ncp1 = NULL,
  ncp0 = NULL,
  dist,
  minimumConditionalError = 0,
  maximumConditionalError = 1,
  ...
)
}
\arguments{
\item{alpha}{The overall type I error rate \eqn{\alpha} of the design. Must be a numeric value between 0 and 1.}

\item{alpha1}{Stage 1 efficacy boundary \eqn{\alpha_1} (p-value scale). Must be a numeric value between 0 and 1. Should be smaller than \code{alpha0}.}

\item{alpha0}{Binding stage 1 futility boundary \eqn{\alpha_0} (p-value scale). Must be a numeric value between 0 and 1. Should be greater than \code{alpha1}. For use of a non-binding futility boundary, specify \code{alpha0=1}.}

\item{conditionalPower}{The target conditional power \eqn{CP} of the design. To guarantee that the function \code{getPsi()} can be solved, the conditional power should be between \code{1-pnorm(2)} and \code{pnorm(2)}. Must be a numeric value.}

\item{ncp1}{Fixed effect assumption, expressed as a non-centrality parameter. Must be a numeric value.}

\item{ncp0}{Minimum value for the interim treatment effect estimate, expressed as a non-centrality parameter. Must be a numeric value.}

\item{dist}{The distribution to be used for the effect size of the likelihood ratio in the optimal conditional error function. Options are \code{"fixed", "normal", "exp", "unif", "maxlr"} for fixed effect size, normally distributed, exponentially distributed, uniformly distributed prior of the effect size and maximum likelihood ratio, respectively.
Each case requires different additional specifications. \cr
\itemize{
\item \code{dist="fixed"} uses one (or more) fixed effect sizes for the likelihood ratio and requires the parameter \code{ncpLR} which provides the non-centrality parameter under which to calculate the likelihood ratio. If \code{ncpLR} contains multiple values, they may be weighted using an additional argument \code{weights}. Omitting \code{weights} automatically leads to equal weighting.
\item \code{dist="normal"} uses a normal prior for the effect size and requires parameters \code{ncpLR} and \code{tau} for the mean and standard deviation of the normal distribution (both on non-centrality parameter scale).
\item \code{dist="exp"} uses an exponential prior for the effect size and requires the parameter \code{kap0} which is the mean of the exponential distribution (on the non-centrality parameter scale).
\item \code{dist="unif"} uses a uniform prior for the effect size and requires the specification of \code{delMax}, which is the maximum of the support for the uniform likelihood ratio distribution (on the non-centrality parameter scale).
\item \code{dist="maxlr"} estimates the non-centrality parameter to be used for the likelihood ratio from the data. No additional parameters must be specified.
}}

\item{minimumConditionalError}{Lower boundary for the optimal conditional error function. Default 0 (no restriction).}

\item{maximumConditionalError}{Upper boundary for the optimal conditional error function. Default value is 1, however, the optimal conditional error function is inherently bounded by the conditional power (assuming no early efficacy stop).}

\item{...}{Additional arguments required for the specification of the likelihood ratio.}
}
\value{
A list of the optimal conditional error design which can be passed on to other package functions.
}
\description{
This function returns a design object in form of a list which contains all important parameters for the optimal CEF.
}
\details{
From the given user specifications, the constant to achieve level condition for control of the overall type I error rate as well as the constants to ensure a non-increasing optimal CEF are automatically calculated.
The level constant is calculated via the helper function \code{getLevelConstant()} and the monotonisation constants are calculated via
the function\code{getMonotonisationConstants()}. See the respective help pages for more details.
}
\examples{
getDesignOptimalConditionalErrorFunction(
alpha=0.025, alpha1=0.001, alpha0=0.5, conditionalPower=0.9,
ncp1=1, dist="fixed", ncpLR=1)

}
