% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInnerPsi.R
\name{getInnerPsi}
\alias{getInnerPsi}
\title{Internal variation of \code{getPsi()} used by \code{findLevelConstant()}}
\usage{
getInnerPsi(firstStagePValue, constant, design)
}
\arguments{
\item{firstStagePValue}{First-stage p-value, internally used as integral argument.}

\item{constant}{Current value of the level constant for the optimal conditional error function, internally used as integral argument.}

\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}
}
\value{
Point value of \code{getPsi()}.
}
\description{
Calculate psi for the given scenario, incl. constant
}
\details{
Internal function called by \code{getIntegral()} that calculates a point value of \code{getPsi()}.
}
\keyword{internal}
