% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInnerPsiConstant.R
\name{getInnerPsiConstant}
\alias{getInnerPsiConstant}
\title{Integration routine called via \code{getIntegralWithConstants()}.}
\usage{
getInnerPsiConstant(firstStagePValue, constant, design)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{constant}{Current value of the level constant for the optimal conditional error function, internally used as integral argument.}

\item{design}{A list created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}
}
\description{
Integration routine called via \code{getIntegralWithConstants()}.
}
