% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntegral.R
\name{getIntegral}
\alias{getIntegral}
\title{Internal calculation of the integral over psi}
\usage{
getIntegral(constant, design)
}
\arguments{
\item{constant}{Current value of the level constant for the optimal conditional error function, internally used as integral argument.}

\item{design}{A list created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}
}
\value{
Distance of integral over psi to (alpha-alpha1).
}
\description{
Helper function that integrates over the optimal conditional error function. Used to find appropriate level constant.
}
\details{
Internal function called by \code{findLevelConstant()} that should be solved (i.e., root should be found).
}
\keyword{internal}
