% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInverseNormalCombinationFunction.R
\name{getInverseNormalCombinationFunction}
\alias{getInverseNormalCombinationFunction}
\title{Calculate Inverse Normal Combination Function}
\usage{
getInverseNormalCombinationFunction(
  firstStagePValue,
  secondStagePValue,
  weights = 1/sqrt(c(2, 2))
)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{secondStagePValue}{p-value of the second stage.}

\item{weights}{Weights assigned to the p-value/test statistic of each stage. Must be a numeric vector assigning a value between 0 and 1 to each stage that sums up to 1.}
}
\value{
Value of the inverse normal combination function.
}
\description{
Calculate the result of the inverse normal combination function.
}
\details{
The inverse normal combination function for stage-wise p-values \eqn{p_1, p_2} is provided in Wassmer & Brannath (2016) as:
\deqn{
C(p_1,p_2) = 1-\Phi(w_1\Phi^{-1}(1-p_1)+w_2\Phi^{-1}(1-p_2)),
}
where \eqn{w_1, w_2} are positive stage-wise weights subject to \eqn{w_1^2+w_2^2=1}.
}
\examples{
getInverseNormalCombinationFunction(
firstStagePValue=0.1, secondStagePValue=0.08)

}
\references{
Wassmer, G. & Brannath, W. (2016). Group sequential and confirmatory adaptive designs in clinical trials. Springer. https://doi.org/https://doi.org/10.1007/978-3-319-32562-0
}
