% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInverseNormalConditionalError.R
\name{getInverseNormalConditionalError}
\alias{getInverseNormalConditionalError}
\title{Calculate Inverse Normal Conditional Error}
\usage{
getInverseNormalConditionalError(
  alpha,
  alpha0,
  alpha1,
  firstStagePValue,
  weights = c(sqrt(0.5), sqrt(0.5)),
  alpha2 = NULL
)
}
\arguments{
\item{alpha}{The overall type I error rate \eqn{\alpha} of the design. Must be a numeric value between 0 and 1.}

\item{alpha0}{Binding stage 1 futility boundary \eqn{\alpha_0} (p-value scale). Must be a numeric value between 0 and 1. Should be greater than \code{alpha1}. For use of a non-binding futility boundary, specify \code{alpha0=1}.}

\item{alpha1}{Stage 1 efficacy boundary \eqn{\alpha_1} (p-value scale). Must be a numeric value between 0 and 1. Should be smaller than \code{alpha0}.}

\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{weights}{Weights assigned to the p-value/test statistic of each stage. Must be a numeric vector assigning a value between 0 and 1 to each stage that sums up to 1.}

\item{alpha2}{Second-stage significance level. Can be specified to speed up calculations, is calculated automatically otherwise. Must be a numeric value between 0 and 1. Default: \code{NULL}.}
}
\value{
Conditional error of the inverse normal combination function.
}
\description{
Calculate the conditional error of the inverse normal method.
}
\details{
The inverse normal conditional error is calculated for a first-stage p-value \eqn{p_1 \in ]\alpha_1, \alpha_0]} as:
\deqn{\alpha_2(p_1) = 1-\Phi(\frac{\Phi^{-1}(1-\alpha_2) - w_1 \Phi^{-1}(1-p_1)}{w_2}),}
where
\itemize{
\item \eqn{\alpha_2} is the second-stage significance level (see \code{getAlpha2InverseNormal()})
\item \eqn{w_1, w_2} are the stage-wise weights
}
}
\examples{
getInverseNormalConditionalError(
alpha = 0.025, alpha0 = 0.5, alpha1 = 0.001, firstStagePValue = 0.09)

}
\references{
Wassmer, G. & Brannath, W. (2016). Group sequential and confirmatory adaptive designs in clinical trials. Springer. https://doi.org/https://doi.org/10.1007/978-3-319-32562-0
}
