% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihoodRatio.R
\name{getLikelihoodRatio}
\alias{getLikelihoodRatio}
\title{Calculate Likelihood Ratio}
\usage{
getLikelihoodRatio(firstStagePValue, design)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{design}{A list created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}
}
\value{
The value of the likelihood ratio for the given specification.
}
\description{
Calculate the likelihood ratio of a p-value for a given distribution.
}
\details{
The calculation of the likelihood ratio for a first-stage p-value \eqn{p_1} is done based on a distributional assumption, specified via the argment \code{dist}.
The different options require different parameters, elaborated in the following.
\itemize{
\item \code{dist="fixed"}: calculates the likelihood ratio for a fixed \eqn{\vartheta}. The non-centrality parameter of the likelihood ratio \eqn{\vartheta} must then be specified via \code{ncpLR} to calculate the likelihood ratio as:
\deqn{l(p_1) = e^{\Phi^{-1}(1-p_1)\vartheta - \vartheta^2/2}}
\item \code{dist="normal"}: calculates the likelihood ratio for a normally distributed prior of \eqn{\vartheta} with mean \code{ncpLR} (\eqn{\vartheta}) and standard deviation \code{tauLR} (\eqn{\tau}). Both parameters must be specified on the non-centrality parameter scale.
\deqn{l(p_1) = (1+\tau^2)^{-\frac{1}{2}}\cdot e^{-(\vartheta/\tau)^2/2 + (\tau\Phi^{-1}(1-p_1) + \vartheta/\tau)^2 / (2\cdot (1+\tau^2))}}
\item \code{dist="exp"}: calculates the likelihood ratio for an exponentially distributed prior of \eqn{\vartheta} with parameter \code{kap0} (\eqn{\kappa}), which is the mean of the exponential distribution, calculated as \eqn{1/\vartheta} as:
\deqn{l(p_1) = \kappa \cdot \sqrt{2\pi} \cdot e^{(\Phi^{-1}(1-p_1)-\kappa)^2/2} \cdot \Phi(\Phi^{-1}(1-p_1)-\kappa)}
\item \code{dist="unif"}: calculates the likelihood ratio for a uniformly distributed prior of \eqn{\vartheta} on the support \eqn{[0, \Delta]}, where \eqn{\Delta} is specified as \code{delMax}.
\deqn{l(p_1) = \frac{\sqrt{2\pi}}{\Delta} \cdot e^{\Phi^{-1}(1-p_1)^2/2} \cdot (\Phi(\Delta - \Phi^{-1}(1-p_1))-p_1)}
\item \code{dist="maxlr"}: the non-centrality parameter \eqn{\vartheta} is estimated from the data and no additional parameters must be specified. The likelihood ratio is estimated from the data as:
\deqn{l(p_1) = e^{max(0, \Phi^{-1}(1-p_1))^2/2}}
The maximum likelihood ratio is always restricted to effect sizes \eqn{\vartheta \geq 0}. (respectively \eqn{p_1 \leq 0.5}).
}
}
\references{
Brannath, W. & Bauer, P. (2004). Optimal conditional error functions for the control of conditional power. Biometrics, 60 (3), 715–723. https://doi.org/10.1111/j.0006-341X.2004.00221.x

Hung, H. M. J., O’Neill, R. T., Bauer, P. & Kohne, K. (1997). The behavior of the p-value when the alternative hypothesis is true. Biometrics. http://www.jstor.org/stable/2533093
}
