% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxInformation.R
\name{getMaxInformation}
\alias{getMaxInformation}
\title{Calculate Maximum Second-stage Information}
\usage{
getMaxInformation(design)
}
\arguments{
\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the optimal conditional error function for the specified case.}
}
\description{
Calculate the maximum information
}
