% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanDifference.R
\name{getMeanDifference}
\alias{getMeanDifference}
\title{Calculate Mean Difference from Non-centrality Parameter}
\usage{
getMeanDifference(
  ncp,
  firstStageSampleSize,
  standardDeviation = 1,
  allocationRatio = 1,
  ncp0 = -Inf
)
}
\arguments{
\item{ncp}{Treatment effect as a non-centrality parameter. Must be a numeric value.}

\item{firstStageSampleSize}{Sample size of the treatment group in the first stage. Must be a numeric value.}

\item{standardDeviation}{Assumed standard deviation (default: 1). Must be a numeric value.}

\item{allocationRatio}{Allocation ratio between the treatment groups, expressed as (Number of active patients)/(Number of control patients). Must be a numeric value.}

\item{ncp0}{Minimum value for the interim treatment effect estimate, expressed as a non-centrality parameter. Must be a numeric value.}
}
\value{
Mean difference for a specified non-centrality parameter.
}
\description{
Calculate the mean difference from specified design parameters.
}
\details{
The calculation of the mean difference \eqn{\Delta} is done as:
\deqn{\Delta = \vartheta \cdot \sigma \cdot \sqrt{\frac{r+1}{n_t}},}
where
\itemize{
\item \eqn{\vartheta} is the effect size, specified as a non-centrality parameter (argument \code{ncp})
\item \eqn{\sigma} is the standard deviation of the effect (equal between groups, argument \code{standardDeviation})
\item \eqn{n_t} is the sample size of the treatment group (argument \code{firstStageSampleSize})
\item \eqn{r} is the allocation ratio between the groups (\eqn{r=n_t/n_c}, argument \code{allocationRatio}).
}
A minimum effect size (non-centrality parameter scale) can be specified via the argument \code{ncp0}.
}
\examples{
getMeanDifference(
ncp = 1.768, firstStageSampleSize = 100)

}
