% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinMaxInformation.R
\name{getMinMaxInformation}
\alias{getMinMaxInformation}
\title{Calculate Expected Second-stage Information}
\usage{
getMinMaxInformation(design, delta_min = NULL, delta_max = NULL)
}
\arguments{
\item{design}{A list created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}

\item{delta_min}{Minimum for treatment effect.}

\item{delta_max}{Maximum for treatment effect.}
}
\description{
Calculate the minimum and maximum information
}
\examples{
design <- getDesignOptimalConditionalErrorFunction(
  alpha = 0.025, alpha1 = 0.001, alpha0 = 0.5,
  conditionalPower = 0.9, delta1Min = 0.2, delta1Max = 0.7,
  firstStageInformation = 9, useInterimEstimate = TRUE,
  likelihoodRatioDistribution = "maxlr"
)

getMinMaxInformation(design, delta_min = 0.2, delta_max = 0.7)

}
