% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonCentralityParameter.R
\name{getNonCentralityParameter}
\alias{getNonCentralityParameter}
\title{Calculate Non-centrality Parameter}
\usage{
getNonCentralityParameter(
  delta,
  firstStageSampleSize,
  standardDeviation = 1,
  allocationRatio = 1,
  delta0 = -Inf
)
}
\arguments{
\item{delta}{Treatment effect, expressed as a mean difference. Must be a numeric value.}

\item{firstStageSampleSize}{Sample size of the treatment group in the first stage. Must be a numeric value.}

\item{standardDeviation}{Assumed standard deviation (default: 1). Must be a numeric value.}

\item{allocationRatio}{Allocation ratio between the treatment groups, expressed as (Number of active patients)/(Number of control patients). Must be a numeric value.}

\item{delta0}{Minimally clinically relevant treatment effect, expressed as a mean difference (default: -Inf). Must be a numeric value.}
}
\value{
(Truncated) Non-centrality parameter.
}
\description{
Calculates the non-centrality parameter for given design parameters.
}
\details{
The non-centrality parameter \eqn{\vartheta} is calculated as
\deqn{\vartheta = \frac{\Delta}{\sigma}\cdot \sqrt{\frac{n_t}{r+1}},}
where
\itemize{
\item \eqn{\Delta} is the effect size, specified as a mean difference (argument \code{delta})
\item \eqn{\sigma} is the standard deviation of the effect (equal between groups, argument \code{standardDeviation})
\item \eqn{n_t} is the sample size of the treatment group (argument \code{firstStageSampleSize})
\item \eqn{r} is the allocation ratio between the groups (\eqn{r=n_t/n_c}, argument \code{allocationRatio}).
}
A minimum effect size (mean difference scale) can be specified via the argument \code{delta0}.
}
\examples{
getNonCentralityParameter(
delta = 0.25, firstStageSampleSize = 100)

}
