% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverallPower.R
\name{getOverallPower}
\alias{getOverallPower}
\title{Calculate the overall power}
\usage{
getOverallPower(design, delta1)
}
\arguments{
\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}

\item{delta1}{The effect size on the mean difference scale at which the overall power should be calculated.}
}
\value{
The overall power of the design at the provided effect size.
}
\description{
Calculate the overall power
}
