% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSecondStageInformation.R
\name{getSecondStageInformation}
\alias{getSecondStageInformation}
\title{Calculate the Second-stage Information}
\usage{
getSecondStageInformation(conditionalError, conditionalPower, delta1)
}
\arguments{
\item{conditionalError}{Conditional error rate of the design.}

\item{conditionalPower}{The target conditional power \eqn{CP} of the design. To guarantee that the function \code{getPsi()} can be solved, the conditional power should be between \code{1-pnorm(2)} and \code{pnorm(2)}. Must be a numeric value.}

\item{delta1}{The effect size to power for.}
}
\value{
The second-stage information.
}
\description{
Calculate second-stage information for given alpha, conditional power and effect size
}
\details{
The second-stage information \eqn{I_{2}} is calculated given a first-stage p-value \eqn{p_1} as:
\deqn{I_{2}(p_1) = \frac{(\Phi^{-1}(1-\alpha_2(p_1)) + \Phi^{-1}(CP))^2}{\Delta_1^2} = \frac{\nu(\alpha_2(p_1))}{\Delta_1^2},}
where
\itemize{
\item \eqn{\alpha_2(p_1)} is the second-stage significance level (e.g., a conditional error)
\item \eqn{CP} is the target conditional power
\item \eqn{\Delta_1} is the assumed treatment effect (expressed as a mean difference).
}
}
\examples{
getSecondStageInformation(
conditionalError = 0.05, conditionalPower = 0.9, delta1 = 0.25)


}
\references{
Brannath, W. & Bauer, P. (2004). Optimal conditional error functions for the control of conditional power. Biometrics, 60 (3), 715–723. https://doi.org/10.1111/j.0006-341X.2004.00221.x
}
