% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSecondStageSampleSize.R
\name{getSecondStageSampleSize}
\alias{getSecondStageSampleSize}
\title{Calculate the Second-stage Sample Size}
\usage{
getSecondStageSampleSize(
  conditionalError,
  conditionalPower,
  delta1,
  allocationRatio = 1,
  standardDeviation = 1
)
}
\arguments{
\item{conditionalError}{Conditional error rate of the design.}

\item{conditionalPower}{The target conditional power \eqn{CP} of the design. To guarantee that the function \code{getPsi()} can be solved, the conditional power should be between \code{1-pnorm(2)} and \code{pnorm(2)}. Must be a numeric value.}

\item{delta1}{Effect size to power for.}

\item{allocationRatio}{Allocation ratio between the treatment groups, expressed as (Number of active patients)/(Number of control patients). Must be a numeric value.}

\item{standardDeviation}{Assumed standard deviation (default: 1). Must be a numeric value.}
}
\value{
The second-stage sample size of the treatment group.
}
\description{
Calculate second-stage sample size for given alpha, conditional power and effect size
}
\details{
The second-stage sample size \eqn{n_{2,t}} of the treatment group is calculated given a first-stage p-value \eqn{p_1} as:
\deqn{n_{2,t}(p_1) = \frac{\sigma^2\cdot(r+1)\cdot (\Phi^{-1}(1-\alpha_2(p_1)) + \Phi^{-1}(CP))^2}{\Delta_1^2} = \frac{\sigma^2\cdot(r+1) \cdot \nu(\alpha_2(p_1))}{\Delta_1^2},}
where
\itemize{
\item \eqn{\sigma} is the assumed (shared) standard deviation of the treatment effect
\item \eqn{r} is the allocation ratio between treatment and control group (\eqn{r=n_t/n_c})
\item \eqn{\alpha_2(p_1)} is the second-stage significance level (e.g., a conditional error)
\item \eqn{CP} is the target conditional power
\item \eqn{\Delta_1} is the assumed treatment effect (expressed as a mean difference).
}
}
\examples{
getSecondStageSampleSize(
conditionalError = 0.05, conditionalPower = 0.9, delta1 = 0.25)


}
\references{
Brannath, W. & Bauer, P. (2004). Optimal conditional error functions for the control of conditional power. Biometrics, 60 (3), 715–723. https://doi.org/10.1111/j.0006-341X.2004.00221.x
}
