% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimulationResults.R
\name{getSimulationResults}
\alias{getSimulationResults}
\title{Simulate trials}
\usage{
getSimulationResults(
  design,
  maxNumberOfIterations = 10000,
  alternative,
  seed = NULL
)
}
\arguments{
\item{design}{A list created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}

\item{maxNumberOfIterations}{Number of trials to be simulated.}

\item{alternative}{Assumed relative effect size.}
}
\value{
Proportion of rejections for the given scenario
}
\description{
Simulate the rejection probability for a given design and alternative.
}
