% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PowerResultsOptimalConditionalError}
\alias{PowerResultsOptimalConditionalError}
\title{Power results for optimal conditional error design}
\description{
A class for power results of the optimal conditional error function.
}

