% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.PowerResultsOptimalConditionalError}
\alias{print.PowerResultsOptimalConditionalError}
\title{Print power results}
\usage{
\method{print}{PowerResultsOptimalConditionalError}(x, ...)
}
\arguments{
\item{x}{Power results object of class \code{PowerResultsOptimalConditionalError}}

\item{...}{Additional arguments required for generic compatibility}
}
\value{
No return value, prints the power calculation results.
}
\description{
Print an overview of exact power results.
}
