% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.SimulationResultsOptimalConditionalError}
\alias{print.SimulationResultsOptimalConditionalError}
\title{Print simulation results}
\usage{
\method{print}{SimulationResultsOptimalConditionalError}(x, ...)
}
\arguments{
\item{x}{Simulation results object of class \code{SimulationResultsOptimalConditionalError}}

\item{...}{Additional arguments}
}
\description{
Print an overview of simulation results.
}
