% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rangeCheck}
\alias{rangeCheck}
\title{Simple range check for numeric variables}
\usage{
rangeCheck(variable, range, allowedEqual, hint = "")
}
\arguments{
\item{variable}{The (named) variable to be checked.}

\item{range}{A vector of length 2 giving the minimum and maximum of the allowed range.}

\item{allowedEqual}{Logical. Are the borders of range valid values?}

\item{hint}{Additional message that may be printed after the error.}
}
\description{
This function performs very basic range checks for numeric variables and throws
an error if the range is violated. A custom hint may be added to the message.
}
