% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SimulationResultsOptimalConditionalError}
\alias{SimulationResultsOptimalConditionalError}
\title{Simulation results for optimal conditional error design}
\description{
A class for simulation results of the optimal conditional error function.
}

