% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary.TrialDesignOptimalConditionalError}
\alias{summary.TrialDesignOptimalConditionalError}
\title{Summary of the optimal conditional error trial design}
\usage{
\method{summary}{TrialDesignOptimalConditionalError}(object, ...)
}
\arguments{
\item{object}{Design object of class \code{TrialDesignOptimalConditionalError}}

\item{...}{Additional arguments required for generic compatibility}
}
\description{
Provide an overview of the optimal conditional error trial design.
}
