% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{TrialDesignOptimalConditionalError}
\alias{TrialDesignOptimalConditionalError}
\title{Optimal Conditional Error Design}
\description{
A class for a trial design object using the optimal conditional error function.
}
\details{
This object should not be created directly; use \code{getDesignOptimalConditionalErrorFunction()} with suitable arguments to create a design.
}

