% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstraintC_max.R
\name{getConstraintC_max}
\alias{getConstraintC_max}
\title{Calculates the upper constraint.}
\usage{
getConstraintC_max(firstStagePValue, design)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the optimal conditional error function for the specified case.}
}
\value{
Value of the upper constraint
}
\description{
Calculates the upper constraint.
}
\details{
Internal function that calculates the upper constraint.
}
