% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConstraintC_min.R
\name{getConstraintC_min}
\alias{getConstraintC_min}
\title{Calculates the lower constraint.}
\usage{
getConstraintC_min(firstStagePValue, design)
}
\arguments{
\item{firstStagePValue}{First-stage p-value or p-values. Must be a numeric vector between 0 and 1.}

\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the optimal conditional error function for the specified case.}
}
\value{
Value of the lower constraint
}
\description{
Calculates the lower constraint.
}
\details{
Internal function that calculates the lower constraint.
}
