% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIntegralWithConstants.R
\name{getIntegralWithConstants}
\alias{getIntegralWithConstants}
\title{Calculate the integral over a partially constant function}
\usage{
getIntegralWithConstants(constant, design)
}
\arguments{
\item{constant}{Current value of the level constant for the optimal conditional error function, internally used as integral argument.}

\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the optimal conditional error function for the specified case.}
}
\value{
Integral over the partially constant function.
}
\description{
This function aims to provide a more accurate integration routine than the \code{integrate()} function
for a partially constant function.
}
\keyword{internal}
