% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLevelConstant.R
\name{getLevelConstant}
\alias{getLevelConstant}
\title{Get Level Constant for Optimal Conditional Error Function}
\usage{
getLevelConstant(design, constant_min = 0, constant_max = 10)
}
\arguments{
\item{design}{A list created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}

\item{constant_min}{Minimum value of the interval on which the constant should be searched. Default = 0.}

\item{constant_max}{Maximum value of the interval on which the constant should be searched. Default = 10.}
}
\value{
A list that contains the constant (element \code{$root}) and other components provided by \code{uniroot()}.
}
\description{
Find the constant required such that the conditional error function meets the overall level condition.
}
\details{
The level condition is defined as:
\deqn{\alpha = \alpha_1 + \int_{\alpha_1}^{\alpha_0} \alpha_2(p_1)dp_1.}
The constant \eqn{c_0} of the optimal conditional error function is calibrated such that it meets the level condition.
For a valid design, the additional following condition must be met to be able to exhaust the level \eqn{\alpha}:
\deqn{\alpha_1 + CP(\alpha_0-\alpha_1)>\alpha.}
This condition is checked by \code{getLevelConstant()} and the execution is terminated if it is not met. \cr
The argument \code{monotonisationConstants} can be calculated beforehand with \code{getMonotonisationConstants()} and be provided
to calculate the level constant for the non-decreasing transformation of the optimal conditional error function. If this argument is omitted,
the constant is calculated for the optimal conditional error function with no transformations.
}
\references{
Brannath, W., & Bauer, P. (2004). Optimal conditional error functions for the control of conditional power. Biometrics, 60 (3), 715–723. https://doi.org/10.1111/j.0006-341X.2004.00221.x

Brannath, W., & Dreher, M. (2024). Optimal monotone conditional error functions. https://arxiv.org/abs/2402.00814
}
