% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinMaxInformation.R
\name{getMinMaxInformation}
\alias{getMinMaxInformation}
\title{Calculate Expected Second-stage Information}
\usage{
getMinMaxInformation(design, delta_min = 0.001, delta_max = NULL)
}
\description{
Calculate the minimum and maximum information
}
