% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMinMaxInformation.R
\name{getMinMaxInformation}
\alias{getMinMaxInformation}
\title{Calculate Expected Second-stage Information}
\usage{
getMinMaxInformation(design, delta_min = NULL, delta_max = NULL)
}
\arguments{
\item{design}{An object of class \code{TrialDesignOptimalConditionalError} created by \code{getDesignOptimalConditionalErrorFunction()}. Contains all necessary arguments to calculate the non-increasing optimal conditional error function for the specified case.}

\item{delta_min}{Minimum for treatment effect.}

\item{delta_max}{Maximum for treatment effect.}
}
\description{
Calculate the minimum and maximum information
}
