% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNuPrime.R
\name{getNuPrime}
\alias{getNuPrime}
\title{Calculate the Derivate of Nu}
\usage{
getNuPrime(alpha, conditionalPower)
}
\arguments{
\item{alpha}{The (conditional) type I error rate of the design. Must be a numeric vector with values between 0 and 1.}

\item{conditionalPower}{The target conditional power \eqn{CP} of the design. To guarantee that the function \code{getPsi()} can be solved, the conditional power should be between \code{1-pnorm(2)} and \code{pnorm(2)}. Must be a numeric value.}
}
\value{
Value for nu prime.
}
\description{
Calculates the derivative of nu for a given conditional error and conditional power.
}
\details{
The function \eqn{\nu'} is defined as
\deqn{\nu'(p_1) = -2 \cdot (\Phi^{-1}(1-\alpha_2(p_1)) + \Phi^{-1}(CP))/\phi(\Phi^{-1}(1-\alpha_2(p_1))).}
Note that in this implementation, the the factor -2 is used instead of -4, which is used in by Brannath & Bauer (2004), as the allocation ratio is not explicitly fixed at r=1.
}
\examples{
getNuPrime(alpha = 0.05, conditionalPower = 0.9)

}
\references{
Brannath, W. & Bauer, P. (2004). Optimal conditional error functions for the control of conditional power. Biometrics, 60 (3), 715–723. https://doi.org/10.1111/j.0006-341X.2004.00221.x
}
