% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.TrialDesignOptimalConditionalError}
\alias{print.TrialDesignOptimalConditionalError}
\title{Print optimal conditional error trial design}
\usage{
\method{print}{TrialDesignOptimalConditionalError}(x, ...)
}
\arguments{
\item{x}{Design object of class \code{TrialDesignOptimalConditionalError}}

\item{...}{Additional arguments required for generic compatibility}
}
\value{
No return value, prints the design.
}
\description{
Print an overview of the specified design parameters.
}
